WEBAPI/requirements.txt
asgiref==3.6.0
astroid==2.15.4
colorama==0.4.6
dill==0.3.6
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.2.0
pluggy==1.0.0
PyJWT==2.6.0
pylint==2.17.3
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.7
typing==3.7.4.3
tzdata==2023.3
wrapt==1.15.0
psycopg2-binary==2.9.6
